servidor/requirements.txt
jsonschema==3.2.0
kafka-python==2.0.2